Enable Adding New Note Feature
Pressing the 'Add Note' will now open AddNoteView where user can add a
new note and save it.

The List View in ContentView now shows the count of the total notes.

Part of save() method is commented as a temporary measure to prevent
permanently storing the notes object in Persistent Stores.